Fixes error handling and returns for nq/dq functions.
Must be able to indicate to the caller whether there was an error during
execution of dq, or whether there was no error but also no messages.

These changes make sure that the return value is false for all errors,
and true when there were no errors, even if there were no messages.

When there are no messages, the return value is true and the msg is
NULL.